+++ internal/meta/meta.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"os"
+	"slices"
 	"strconv"
 
 	"google.golang.org/grpc/metadata"
@@ -55,7 +56,7 @@ func WithRequestTypeOption(requestType string) Option {
 
 func AllowOption(feature string) Option {
 	return func(m *Meta) {
-		m.capabilities = append(m.capabilities, feature)
+		m.capabilities = slices.Compact(append(m.capabilities, feature))
 	}
 }