+++ sql.go
@@ -6,7 +6,6 @@ import (
 	"database/sql/driver"
 	"fmt"
 
-	"github.com/ydb-platform/ydb-go-sdk/v3/config"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/bind"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/tx"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
@@ -219,25 +218,6 @@ func WithDatabaseSQLTrace(
 	opts ...trace.DatabaseSQLComposeOption,
 ) ConnectorOption {
 	return xsql.WithTrace(&t, opts...)
-}
-
-type uoption Option
-
-func (uoption) Apply(c *xsql.Connector) error {
-	return xsql.WithDisableServerBalancer().Apply(c)
-}
-
-var hintSessionBalancerOption = uoption(func(ctx context.Context, d *Driver) error {
-	d.options = append(d.options, config.WithDisableSessionBalancer())
-
-	return nil
-})
-
-// WithDisableServerBalancer returns a ConnectorOption that disables session balancing for the SQL connector.
-// If you want to use it as an Option for the native driver, wrap it with `ydb.Option`.
-func WithDisableServerBalancer() uoption {
-	// Separate type is necessary so that the option works for both the SQL driver and the native driver
-	return hintSessionBalancerOption
 }
 
 type SQLConnector interface {